+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added equivalent years of record, drainage area export, and order index to equations table 
 - Allow user to sort Variable Types by Statistic Group in Settings
 - Allow user to add unit types and statistic groups to regression types
+- Batch upload option for variables
 
 ### Changed